# perf: only log dependencies