--- a/classport-instr-agent/src/main/java/io/github/chains_project/classport/instrumentation/Agent.java
+++ b/classport-instr-agent/src/main/java/io/github/chains_project/classport/instrumentation/Agent.java
@@ -1,5 +1,15 @@
 package io.github.chains_project.classport.instrumentation;
 
+import io.github.chains_project.classport.commons.AnnotationReader;
+import io.github.chains_project.classport.commons.ClassportInfo;
+import io.github.chains_project.classport.instrumentation.granularity.ClassInvocation;
+import io.github.chains_project.classport.instrumentation.granularity.Granularity;
+import io.github.chains_project.classport.instrumentation.granularity.MethodInvocation;
+import io.github.chains_project.classport.instrumentation.granularity.RecordingStrategy;
+import org.objectweb.asm.ClassReader;
+import org.objectweb.asm.ClassVisitor;
+import org.objectweb.asm.ClassWriter;
+
 import java.lang.instrument.ClassFileTransformer;
 import java.lang.instrument.Instrumentation;
 import java.nio.file.Path;
@@ -12,81 +22,80 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import io.github.chains_project.classport.instrumentation.granularity.Granularity;
-import org.objectweb.asm.ClassReader;
-import org.objectweb.asm.ClassVisitor;
-import org.objectweb.asm.ClassWriter;
-
-import io.github.chains_project.classport.commons.AnnotationReader;
-import io.github.chains_project.classport.commons.ClassportInfo;
-
 public class Agent {
-    private static final Map<String, ClassportInfo> annotationCache = new ConcurrentHashMap<>();
-    private static final String TIMESTAMP = new SimpleDateFormat("yyyyMMddHHmmss").format(new Date());
-    static String OUTPUT_FILE = "_" + TIMESTAMP + ".csv";
-    static Path OUTPUT_PATH_DIR = Paths.get(System.getProperty("user.dir"), "output");
-    static List<String> nonAnnotatedClasses = new ArrayList<>();
-    static Granularity granularity = Granularity.METHOD; // Default
+	private static final Map<String, ClassportInfo> annotationCache = new ConcurrentHashMap<>();
+	private static final String TIMESTAMP = new SimpleDateFormat("yyyyMMddHHmmss").format(new Date());
+	static String OUTPUT_FILE = "_" + TIMESTAMP + ".csv";
+	static Path OUTPUT_PATH_DIR = Paths.get(System.getProperty("user.dir"), "output");
+	static List<String> nonAnnotatedClasses = new ArrayList<>();
+	static Granularity granularity = Granularity.METHOD; // Default
+	private static RecordingStrategy recordingStrategy;
 
+	public static void premain(String agentArgs, Instrumentation inst) {
+		System.out.println("Agent started");
+		if (agentArgs != null && !agentArgs.isEmpty()) {
+			String[] args = agentArgs.split(",");
+			if (args.length > 0) {
+				OUTPUT_FILE = args[0].concat(OUTPUT_FILE);
+			}
+			if (args.length > 1) {
+				OUTPUT_PATH_DIR = Paths.get(args[1]);
+			} else {
+				OUTPUT_PATH_DIR = Paths.get(System.getProperty("user.dir"), "output");
+			}
+			if (args.length > 2) {
+				granularity = Granularity.fromString(args[2].toUpperCase());
+			}
+			System.out.println("Output file: " + OUTPUT_FILE);
+			System.out.println("Output path: " + OUTPUT_PATH_DIR);
+			System.err.println("Not annotated classes will be saved in " + OUTPUT_FILE.replace(".csv", "_nonAnnotatedClasses.txt"));
+			Path outputPath = OUTPUT_PATH_DIR.resolve(OUTPUT_FILE);
 
+			recordingStrategy = switch (granularity) {
+				case CLASS -> new ClassInvocation(outputPath);
+				case METHOD -> new MethodInvocation(outputPath);
+			};
+			recordingStrategy.initializeCSVHeader(outputPath);
+			Runtime.getRuntime().addShutdownHook(new Thread(recordingStrategy::writeToFile));
 
-    public static void premain(String agentArgs, Instrumentation inst) {
-        System.out.println("Agent started");
-        if (agentArgs != null && !agentArgs.isEmpty()) {
-            String[] args = agentArgs.split(",");
-            if (args.length > 0) {
-                OUTPUT_FILE = args[0].concat(OUTPUT_FILE);
-            }
-            if (args.length > 1) {
-                OUTPUT_PATH_DIR = Paths.get(args[1]);
-            } else {
-                OUTPUT_PATH_DIR = Paths.get(System.getProperty("user.dir"), "output");
-            }
-            if (args.length > 2) {
-                granularity = Granularity.fromString(args[2].toUpperCase());
-            }
-            System.out.println("Output file: " + OUTPUT_FILE);
-            System.out.println("Output path: " + OUTPUT_PATH_DIR);
-            System.err.println("Not annotated classes will be saved in " + OUTPUT_FILE.replace(".csv", "_nonAnnotatedClasses.txt"));
-        }
-        inst.addTransformer(new MethodTransformer());
-    }
+		}
+		inst.addTransformer(new MethodTransformer());
+	}
 
-    static class MethodTransformer implements ClassFileTransformer {
-        @Override
-        public byte[] transform(ClassLoader loader, String className, Class<?> classBeingRedefined, ProtectionDomain protectionDomain, byte[] classfileBuffer) {
-            if (classBeingRedefined != null ) {
-                return classfileBuffer;
-            }
-            
-            return transformClass(className, classfileBuffer);
-        }
+	static class MethodTransformer implements ClassFileTransformer {
+		@Override
+		public byte[] transform(ClassLoader loader, String className, Class<?> classBeingRedefined, ProtectionDomain protectionDomain, byte[] classfileBuffer) {
+			if (classBeingRedefined != null) {
+				return classfileBuffer;
+			}
 
-        public byte[] transformClass(String className, byte[] classfileBuffer) {
-            try {
-                ClassportInfo annotationInfo = getAnnotationInfo(className, classfileBuffer);
-                if (annotationInfo != null) {
-                    return applyTransformations(classfileBuffer, className, annotationInfo, granularity);
-                } 
-                else {
-                    nonAnnotatedClasses.add(className); 
-                }
-            } catch (Exception e) {
-                System.err.println("Error transforming class " + className + ": " + e.getMessage());
-            }
-            return classfileBuffer;
-        }
+			return transformClass(className, classfileBuffer);
+		}
 
-        public ClassportInfo getAnnotationInfo(String className, byte[] classfileBuffer) {
-            return annotationCache.computeIfAbsent(className, key -> AnnotationReader.getAnnotationValues(classfileBuffer));
-        }
+		public byte[] transformClass(String className, byte[] classfileBuffer) {
+			try {
+				ClassportInfo annotationInfo = getAnnotationInfo(className, classfileBuffer);
+				if (annotationInfo != null) {
+					return applyTransformations(classfileBuffer, className, annotationInfo);
+				} else {
+					nonAnnotatedClasses.add(className);
+				}
+			} catch (Exception e) {
+				System.err.println("Error transforming class " + className + ": " + e.getMessage());
+			}
+			return classfileBuffer;
+		}
 
-        public byte[] applyTransformations(byte[] classfileBuffer, String className, ClassportInfo annotationInfo, Granularity granularity) {
-            ClassWriter writer = new ClassWriter(ClassWriter.COMPUTE_MAXS);
-            ClassReader reader = new ClassReader(classfileBuffer);
-            ClassVisitor visitor = new MethodInterceptorVisitor(writer, className, annotationInfo, OUTPUT_FILE, OUTPUT_PATH_DIR, granularity);
-            reader.accept(visitor, 0);
-            return writer.toByteArray();
-        }
-    }
+		public ClassportInfo getAnnotationInfo(String className, byte[] classfileBuffer) {
+			return annotationCache.computeIfAbsent(className, key -> AnnotationReader.getAnnotationValues(classfileBuffer));
+		}
+
+		public byte[] applyTransformations(byte[] classfileBuffer, String className, ClassportInfo annotationInfo) {
+			ClassWriter writer = new ClassWriter(ClassWriter.COMPUTE_MAXS);
+			ClassReader reader = new ClassReader(classfileBuffer);
+			ClassVisitor visitor = new MethodInterceptorVisitor(writer, className, annotationInfo, OUTPUT_PATH_DIR, recordingStrategy);
+			reader.accept(visitor, 0);
+			return writer.toByteArray();
+		}
+	}
 }