--- a/classport-instr-agent/src/main/java/io/github/chains_project/classport/instrumentation/granularity/ClassInvocation.java
+++ b/classport-instr-agent/src/main/java/io/github/chains_project/classport/instrumentation/granularity/ClassInvocation.java
@@ -17,22 +17,8 @@ public class ClassInvocation implements RecordingStrategy {
 	}
 
 	@Override
-	public void initializeCSVHeader(Path outputPath) {
-		// Write header to the file if it doesn't exist or is empty
-		try (BufferedWriter writer = new BufferedWriter(new FileWriter(outputPath.toFile(), true))) {
-			File file = outputPath.toFile();
-			if (file.length() == 0) {
-				writer.write("Class,sourceProjectId,isDirect,id,artefact,group,version,childIds\n");
-				writer.flush();
-			}
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-	}
-
-	@Override
-	public void addToInvokeLater (String className, String methodName, String classportInfo) {
-		set.add(className + "," + classportInfo);
+	public void addToInvokeLater (String content) {
+		set.add(content);
 	}
 
 	@Override