… classport-instr-agent/src/main/java/io/github/chains_project/classport/instrumentation/granularity/DependencyInvocation.java
… classport-instr-agent/src/main/java/io/github/chains_project/classport/instrumentation/granularity/MethodInvocation.java
… classport-instr-agent/src/main/java/io/github/chains_project/classport/instrumentation/granularity/RecordingStrategy.java
… classport-instr-agent/src/test/java/io/github/chains_project/classport/instrumentation/MethodInterceptorVisitorTest.java
… maven-plugin/src/test/java/io/github/chains_project/classport/plugin/EmbeddingMojoTest.java