Since the reason for performance bottleneck was I/O operation, I added a mode `CLASS` that only records classes.

Performance before on PDFBox `./get_overhead.sh pdfbox` _before_ PR changes
```
Average for agent: 6.743 s
DEBUG: AVG_BASELINE=2.317
DEBUG: AVG_AGENT=6.743
Baseline Average: 2.317s
Agent Average: 6.743s
Overhead: 191.00%
```

Performance before on PDFBox `./get_overhead.sh pdfbox` _after_ PR changes

```
DEBUG: AVG_BASELINE=2.051
DEBUG: AVG_AGENT=3.542
Baseline Average: 2.051s
Agent Average: 3.542s
Overhead: 72.00%
```