@algomaster99 2025-06-30
The performance seems to have improved a bit
```
Without agent: 199.541 ms/op                                                                                                                                      ✔  at 13:52:40 
With agent:    499.030 ms/op
Overhead:      150.09%
```